Create LICENSE
I just copied and pasted from:
https://raw.githubusercontent.com/rackt/react-modal/master/LICENSE

Since it is what seems to be applied to most of the repositories.

Side note -
I first forked the react-router-mega-demo and was going to add a License file and make a pull request, but after making the local commit I realized that I would be publicly posting the code under a license with someone elses name attached.  It had never occurred to me before that making a pull request for a LIcense file would be any different than a piece of code.